SimpleFeatures

Link to GitHub Repo

Working with simple feature GIS data in Julia

Fast reading, analyzing, manipulating, and writing vector GIS data in Julia. This package:

Represents feature data as a DataFrame

Stores coordinate reference system (crs) info with the DataFrame and uses it automatically

Interfaces with GDAL and GEOS (coming soon!) for fast operations

About

SimpleFeatures creates a custom type, a SimpleFeature object, that contains a DataFrame of features, crs info, and geometry type info.

Geometry data within the dataframe are stored as another custom type, an sfgeom object. This sfgeom type contains a geometry's Well-known binary (WKB) representation and a preview of the Well-known text (WKT). See more below in the section "How it works".

Example:

SimpleFeature
---------
geomtype:  wkbPolygon
crs:       PROJCRS["NAD83(2011) / UTM zone 17N",BASEGEOGCRS["NAD83(2011)",DATUM["NAD83 (National Spatial Refere..."
---------
features:  
1000×2 DataFrame
  Row │ geom                          lyr.1 
      │ sfgeom                        Int32 
──────┼─────────────────────────────────────
    1 │ POLYGON ((853787 3905499,...      1
    2 │ POLYGON ((853800 3905499,...      1
    3 │ POLYGON ((853803 3905499,...      1
  ⋮   │              ⋮                  ⋮
  998 │ POLYGON ((904045 3905468,...      1
  999 │ POLYGON ((905355 3905468,...      1
 1000 │ POLYGON ((905561 3905469,...      1
                            994 rows omitted

Functions from this package operate on and return SimpleFeature objects.

To directly access the DataFrame, crs info, or geometry info, just add .df, .crs, or .geomtype at the end of your SimpleFeature object's name.

Installation

SimpleFeatures can be installed directly from GitHub using:

using Pkg
Pkg.add("https://github.com/acgold/SimpleFeatures.jl.git")

Getting Started

Load SimpleFeatures.jl

import SimpleFeatures as sf

Read & write data

Most GIS types can be read and written using the functions st_read and st_write, respectively.

x = sf.st_read("data/test.gpkg")

SimpleFeature
---------
geomtype:  wkbPolygon
crs:       PROJCRS["NAD83(2011) / UTM zone 17N",BASEGEOGCRS["NAD83(2011)",DATUM["NAD83 (National Spatial Refere..."
---------
features:  
1000×2 DataFrame
  Row │ geom                          lyr.1 
      │ sfgeom                        Int32 
──────┼─────────────────────────────────────
    1 │ POLYGON ((853787 3905499,...      1
    2 │ POLYGON ((853800 3905499,...      1
    3 │ POLYGON ((853803 3905499,...      1
  ⋮   │              ⋮                  ⋮
  998 │ POLYGON ((904045 3905468,...      1
  999 │ POLYGON ((905355 3905468,...      1
 1000 │ POLYGON ((905561 3905469,...      1
                            994 rows omitted

sf.st_write("data/new_test.gpkg", x)

"data/new_test.gpkg"

Operations

SimpleFeatures offers some basic functions and will offer more in future releases.

Current functionality is:

Casting geometries to different types: sf.st_cast

Buffering: sf.st_buffer (GDAL. Will be GEOS soon)

Reprojecting: sf.st_transform (GDAL)

Segmentizing a line: sf.st_segmentize (GDAL)

Converting between sfgeom objects and GDAL or GEOS (coming soon): sf.sfgeom_to_gdal, sf.gdal_to_sfgeom

Cast polygons to linestrings

In this example, SimpleFeatures will cast each polygon to a multilinestring and then to a linestrings. Some polygons had holes (multiple lines per polygon), so the resulting DataFrame has more rows than the original. In cases such as this, SimpleFeatures adds a column of the geometry type + "ID" (e.g. _MultiLineStringID) that preserves which multigeometry type the split geometry belonged to.

lines = sf.st_cast(df, "linestring")

SimpleFeature
---------
geomtype:  wkbLineString
crs:       PROJCRS["NAD83(2011) / UTM zone 17N",BASEGEOGCRS["NAD83(2011)",DATUM["NAD83 (National Spatial Refere..."
---------
features:  
1022×3 DataFrame
  Row │ lyr.1  _MultiLineStringID  geom                         
      │ Int32  Int64               sfgeom                       
──────┼─────────────────────────────────────────────────────────
    1 │     1                   1  LINESTRING (853787 390549...
    2 │     1                   2  LINESTRING (853800 390549...
    3 │     1                   3  LINESTRING (853803 390549...
  ⋮   │   ⋮            ⋮                        ⋮
 1020 │     1                 998  LINESTRING (904045 390546...
 1021 │     1                 999  LINESTRING (905355 390546...
 1022 │     1                1000  LINESTRING (905561 390546...
                                               1016 rows omitted

Buffer

Using the linestrings from the st_cast example above, we will add a 10m buffer.

buffered_lines = sf.st_buffer(lines, 10) # buffer distance is in units of the crs. Meters in this example

SimpleFeature
---------
geomtype:  wkbPolygon
crs:       PROJCRS["NAD83(2011) / UTM zone 17N",BASEGEOGCRS["NAD83(2011)",DATUM["NAD83 (National Spatial Refere..."
---------
features:  
1022×3 DataFrame
  Row │ lyr.1  _MultiLineStringID  geom                         
      │ Int32  Int64               sfgeom                       
──────┼─────────────────────────────────────────────────────────
    1 │     1                   1  POLYGON ((853787 3905509,...
    2 │     1                   2  POLYGON ((853800 3905509,...
    3 │     1                   3  POLYGON ((853803 3905509,...
  ⋮   │   ⋮            ⋮                        ⋮
 1020 │     1                 998  POLYGON ((904045 3905478,...
 1021 │     1                 999  POLYGON ((905355 3905478,...
 1022 │     1                1000  POLYGON ((905556.62823740...
                                               1016 rows omitted

Reproject

Let's reproject the polygons we just made with st_buffer.

reprojected_buffer = sf.st_transform(x, GeoFormatTypes.EPSG(5070))

SimpleFeature
---------
geomtype:  wkbPolygon
crs:   5070
---------
features:  
1000×2 DataFrame
  Row │ lyr.1  geom                         
      │ Int32  sfgeom                       
──────┼─────────────────────────────────────
    1 │     1  POLYGON ((1693276.9257186...
    2 │     1  POLYGON ((1693289.637015 ...
    3 │     1  POLYGON ((1693292.5703908...
  ⋮   │   ⋮                 ⋮
  998 │     1  POLYGON ((1742412.4620032...
  999 │     1  POLYGON ((1743692.7581942...
 1000 │     1  POLYGON ((1743893.9246629...
                            994 rows omitted

How it works

GIS data as a DataFrame

SimpleFeatures reads vector GIS data from files and creates a DataFrame containing geometries and their attributes.

Each row is a feature, and geometry data for each feature are stored in the geom column.

Example:

1000×2 DataFrame
  Row │ geom                          lyr.1 
      │ sfgeom                        Int32 
──────┼─────────────────────────────────────
    1 │ POLYGON ((853787 3905499,...      1
    2 │ POLYGON ((853800 3905499,...      1
    3 │ POLYGON ((853803 3905499,...      1
  ⋮   │              ⋮                  ⋮
  998 │ POLYGON ((904045 3905468,...      1
  999 │ POLYGON ((905355 3905468,...      1
 1000 │ POLYGON ((905561 3905469,...      1
                            994 rows omitted

Metadata: crs and geomtype

These metadata (.crs, .geomtype) are stored as a attributes within a SimpleFeature object and can be modified or accessed directly.

Example::

x.crs 

GeoFormatTypes.WellKnownText2{GeoFormatTypes.Unknown}(GeoFormatTypes.Unknown(), "PROJCRS[\"NAD83(2011) / UTM zone 17N\",BASEGEOGCRS[\"NAD83(2011)\",DATUM[...

x.geomtype

wkbPolygon::OGRwkbGeometryType = 3

Types:

crs: GeoFormatTypes.GeoFormat

geomtype: ArchGDAL.OGRwkbGeometryType

Geometries as sfgeom objects

Each row of the geom column is an sfgeom object that has two attributes:

wkb: A vector (type Vector{UInt8}) of the Well-known binary (WKB) representation of the geometry.

Example: julia 93-element Vector{UInt8}: 0x01 0x03 0x00 0x00 0x00 ⋮ 0x80 0xed 0xcb 0x4d 0x41

preview: A string that shows an abbreviated Well-known text (WKT) string.

Example:

julia POLYGON ((853787 3905499,...

This data structure allows SimpleFeatures to:

interface with GDAL and GEOS (coming soon) for fast operations (using using ArchGDAL.jl and LibGEOS.jl, respectively)

provide a preview of the geometry's WKT for viewing

represent geometries (wkb) in Julia without needing to define a new SimpleFeatures type for each type of geometry. SimpleFeatures uses ArchGDAL geometry types for geometry type info.

This package builds on existing Julia GIS packages GeoDataFrames and ArchGDAL.

Long-term Goals

To provide full support for simple features access

To provide an intuitive workflow for reading, analyzing, manipulating, and writing vector GIS data in Julia

Inspired by the R package sf. This package is in Alpha and will change significantly in the future.

Reference

SimpleFeatures.st_area

SimpleFeatures.st_buffer

SimpleFeatures.st_cast

SimpleFeatures.st_segmentize

SimpleFeatures.st_transform

SimpleFeatures.st_write

SimpleFeatures.toProjJSON

SimpleFeatures.toWKT2

SimpleFeatures.st_area — Method

st_area(x::SimpleFeature; geom_columns=(:geom,))

Returns a vector of geometry areas in units of the crs. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_buffer — Method

st_buffer(x::DataFrame, d::Number; geom_columns=(:geom,))

Create a new DataFrame that is buffered by the provided distance d in units of the crs. The resulting object is stored in memory as a GeoPackage by default, but a filename fn can be provided. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_cast — Method

st_cast(x::SimpleFeature, to::String)

Cast features geometries to the requested type. See below for the ordered list of to values. Any decomposition from a multigeometry type to a single geometry type will add a unique ID of the multigeometry object as a column (e.g., _MultiPolygonID) so the original multigeometry can be re-created later.

Hierarchy of to values:

"multipolygon"

"polygon"

"multilinestring"

"linestring"

"multipoint"

"point"

SimpleFeatures.st_segmentize — Method

st_segmentize(x::DataFrame, max_length::Number; geom_columns=(:geom,))

Create a new DataFrame that contains LineString geometries that have been sliced into lines of max_length. The resulting object is stored in memory as a GeoPackage by default, but a filename fn can be provided. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_transform — Method

st_transform(x::DataFrame, crs::GFT.GeoFormat; fn::Union{Nothing, AbstractString}=nothing, layer_name::AbstractString="data", driver::Union{Nothing,AbstractString}="GPKG", options::Dict{String,String}=Dict{String,String}(), geom_columns=(:geom,))

Create a new DataFrame that is projected to the provided crs. The resulting object is stored in memory as a GeoPackage by default, but a filename fn can be provided. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_write — Method

st_write(fn::AbstractString, table; layer_name="data", geom_column=:geometry, crs::Union{GFT.GeoFormat,Nothing}=nothing, driver::Union{Nothing,AbstractString}=nothing, options::Vector{AbstractString}=[], geom_columns::Set{Symbol}=(:geometry))

Write the provided table to fn. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.toProjJSON — Method

toProjJSON(spref::AbstractSpatialRef)

Convert this SRS into ProjJSON format.

SimpleFeatures.toWKT2 — Method

toWKT2(spref::AbstractSpatialRef)

Convert this SRS into WKT2 format.