Reference

SimpleFeatures.df_to_sf

SimpleFeatures.sf_to_df

SimpleFeatures.st_area

SimpleFeatures.st_bbox

SimpleFeatures.st_buffer

SimpleFeatures.st_cast

SimpleFeatures.st_centroid

SimpleFeatures.st_combine

SimpleFeatures.st_difference

SimpleFeatures.st_disjoint

SimpleFeatures.st_distance

SimpleFeatures.st_intersection

SimpleFeatures.st_intersects

SimpleFeatures.st_length

SimpleFeatures.st_nearest_points

SimpleFeatures.st_segmentize

SimpleFeatures.st_simplify

SimpleFeatures.st_transform

SimpleFeatures.st_union

SimpleFeatures.st_write

SimpleFeatures.toProjJSON

SimpleFeatures.toWKT2

SimpleFeatures.df_to_sf — Function

dftosf(x::DataFrame, crs::GFT.GeoFormat=GFT.EPSG(4326); geom_column=:geom)

Convert a DataFrame containing a column of ArchGDAL geometries to a new SimpleFeature object.

SimpleFeatures.sf_to_df — Method

sftodf(x::SimpleFeature; geom_column=:geom)

Convert a SimpleFeature object to a DataFrame containing a column of ArchGDAL geometries.

SimpleFeatures.st_area — Method

st_area(x::SimpleFeature; geom_column=:geom)

Returns a vector of geometry areas in units of the crs.

SimpleFeatures.st_bbox — Method

st_bbox(x::SimpleFeature; geom_column=:geom)

Create a bounding box polygon for the provided SimpleFeature object. Resulting polygon is and AG.wkbPolygon type.

SimpleFeatures.st_buffer — Method

st_buffer(x::SimpleFeature, d::Number; geom_column=:geom)
st_buffer(x::SimpleFeature, d::String; geom_column=:geom)

Create a new SimpleFeature object that is buffered by the provided distance d in units of the crs. d can be a number or a string representing the column of the SimpleFeature DataFrame that contains numbers to use for the buffer distance.

SimpleFeatures.st_cast — Method

st_cast(x::SimpleFeature, to::String; groupid::Union{String,Nothing}=nothing, geom_column=:geom, kwargs...)

Cast features geometries to the requested type (to). See below for the ordered list of to values. Supply a groupid for aggregating geometries by a variable. Any decomposition from a multigeometry type to a single geometry type will add a unique ID of the multigeometry object as a column (e.g., _MultiPolygonID) so the original multigeometry can be re-created later.

Hierarchy of to values:

"multipolygon"

"polygon"

"multilinestring"

"linestring"

"multipoint"

"point"

SimpleFeatures.st_centroid — Method

st_centroid(x::SimpleFeature; geom_column=:geom)

Create a SimpleFeature object of point centroids of the input geometries.

SimpleFeatures.st_combine — Method

st_combine(x::SimpleFeature; geom_column=:geom)

Create a new SimpleFeature object that has combined all features into the relevant multigeometry type. No attributes of the original object are returned.

SimpleFeatures.st_difference — Method

st_difference(x::SimpleFeature, y::SimpleFeature; geom_column=:geom)

Create a new SimpleFeature object that is the difference of x and y in the geometry type of x.

SimpleFeatures.st_disjoint — Method

st_disjoint(x::SimpleFeature, y::SimpleFeature; geom_column=:geom, sparse::Bool=true)

Returns a sparse index list of disjoint features.

SimpleFeatures.st_distance — Method

st_distance(x::SimpleFeature, y::SimpleFeature; geom_column=:geom)

Returns a matrix of distances between each x and y geometry in units of the crs.

SimpleFeatures.st_intersection — Method

st_intersection(x::SimpleFeature, y::SimpleFeature; geom_column=:geom)

Create a new SimpleFeature object that is the intersection of x and y in the geometry type of y.

SimpleFeatures.st_intersects — Method

st_intersects(x::SimpleFeature, y::SimpleFeature; geom_column=:geom, sparse::Bool=true)

Returns a sparse index list of intersecting features.

SimpleFeatures.st_length — Method

st_length(x::SimpleFeature; geom_column=:geom)

Returns a vector of geometry lengths in units of the crs.

SimpleFeatures.st_nearest_points — Method

st_nearest_points(x::SimpleFeature, y::SimpleFeature; geom_column=:geom)

Returns a SimpleFeature object of lines between nearest points of each feature in x to y and columns indicating which row of x and y each line refers to.

SimpleFeatures.st_segmentize — Method

st_segmentize(x::SimpleFeature, max_length::Number; geom_column=:geom)

Create a new SimpleFeature object whose segments have been sliced into segments no larger than the provided max_length in units of the crs. See ArchGDAL.segmentize! for more info

SimpleFeatures.st_simplify — Function

st_simplify(x::SimpleFeature; geom_column=:geom)

Create a SimpleFeature object with simplified input geometries.

SimpleFeatures.st_transform — Method

st_transform(x::SimpleFeature, crs::GFT.GeoFormat; geom_column=:geom, order=:compliant)

Create a new SimpleFeature object that is projected to the provided crs.

SimpleFeatures.st_union — Method

st_union(x::SimpleFeature; geom_column=:geom)

Create a SimpleFeature object that unions all input geometries.

SimpleFeatures.st_write — Method

st_write(fn::AbstractString, table; layer_name="data", geom_column=:geometry, crs::Union{GFT.GeoFormat,Nothing}=nothing, driver::Union{Nothing,AbstractString}=nothing, options::Vector{AbstractString}=[], geom_columns::Set{Symbol}=(:geometry))

Write the provided table to fn. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.toProjJSON — Method

toProjJSON(spref::AbstractSpatialRef)

Convert this SRS into ProjJSON format.

SimpleFeatures.toWKT2 — Method

toWKT2(spref::AbstractSpatialRef)

Convert this SRS into WKT2 format.