SimpleFeatures

Documentation for SimpleFeatures.jl.

Working with simple feature GIS data in Julia.

Builds on existing Julia GIS packages GeoDataFrames.jl and ArchGDAL.jl to:

Handle operations as a DataFrame (i.e., DataFrame in, DataFrame out)

Store coordinate reference system info with the DataFrame and use it automatically

Create new copies of underlying geographic data (GDAL) when performing functions (unless specified with a !)

Long-term Goals

To provide full support for simple features access

To provide an intuitive workflow for reading, analyzing, manipulating, and writing vector GIS data in Julia

Inspired by the R package sf. This package is very much in Alpha and is currently intended for my personal use.

SimpleFeatures.st_buffer

SimpleFeatures.st_copy

SimpleFeatures.st_crs

SimpleFeatures.st_is_spdf

SimpleFeatures.st_read

SimpleFeatures.st_set_crs

SimpleFeatures.st_transform

SimpleFeatures.st_write

SimpleFeatures.toProjJSON

SimpleFeatures.toWKT2

SimpleFeatures.st_buffer — Method

st_buffer(x::DataFrame, d::Number; fn::Union{Nothing, AbstractString}=nothing, layer_name::AbstractString="data", driver::Union{Nothing,AbstractString}="GPKG", options::Dict{String,String}=Dict{String,String}(), geom_columns=(:geom,))

Create a new DataFrame that is buffered by the provided distance d in units of the crs. The resulting object is stored in memory as a GeoPackage by default, but a filename fn can be provided. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_copy — Method

st_copy(fn::AbstractString, table; layer_name="data", geom_column=:geometry, crs::Union{GFT.GeoFormat,Nothing}=nothing, driver::Union{Nothing,AbstractString}=nothing, options::Vector{AbstractString}=[], geom_columns::Set{Symbol}=(:geometry))

Copy the provided table to fn. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_crs — Method

st_crs(x::DataFrame) Extract the crs object from the DataFrame's metadata

SimpleFeatures.st_is_spdf — Method

stisspdf(x::DataFrame) Return if DataFrame contains crs metadata and geometry column (true) or is missing one or both items.

SimpleFeatures.st_read — Method

st_read(fn::AbstractString; kwargs...)
st_read(fn::AbstractString, layer::Union{Integer,AbstractString}; kwargs...)

Read a file into a DataFrame. Any kwargs are passed onto ArchGDAL here. By default you only get the first layer, unless you specify either the index (0 based) or name (string) of the layer.

SimpleFeatures.st_set_crs — Method

stsetcrs(x::DataFrame) Set the crs object within the DataFrame's metadata. Metadata will be created if it does not exist. This does not do any projection and will overwrite any existing crs info. To project to a different crs, see st_transform

SimpleFeatures.st_transform — Method

st_transform(x::DataFrame, crs::GFT.GeoFormat; fn::Union{Nothing, AbstractString}=nothing, layer_name::AbstractString="data", driver::Union{Nothing,AbstractString}="GPKG", options::Dict{String,String}=Dict{String,String}(), geom_columns=(:geom,))

Create a new DataFrame that is projected to the provided crs. The resulting object is stored in memory as a GeoPackage by default, but a filename fn can be provided. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.st_write — Method

st_write(fn::AbstractString, table; layer_name="data", geom_column=:geometry, crs::Union{GFT.GeoFormat,Nothing}=nothing, driver::Union{Nothing,AbstractString}=nothing, options::Vector{AbstractString}=[], geom_columns::Set{Symbol}=(:geometry))

Write the provided table to fn. The geom_column is expected to hold ArchGDAL geometries.

SimpleFeatures.toProjJSON — Method

toProjJSON(spref::AbstractSpatialRef)

Convert this SRS into ProjJSON format.

SimpleFeatures.toWKT2 — Method

toWKT2(spref::AbstractSpatialRef)

Convert this SRS into WKT2 format.